Add simple Node Exporter dashboard
This is very losely based on https://grafana.com/grafana/dashboards/6014, but
we don't have all necessary features yet (like different legend strings for
different queries in the same graph).

This also adds a quick URL parameter hack to be able to choose from one of
multiple checked-in sample-data dashboards.

Signed-off-by: Julius Volz <julius.volz@gmail.com>