[IGNORE] Reduce duplicate github actions (#884)
Prior to this change, the github actions in `ci.yaml` would run on
pull request and push events. This means that when maintainers work in
branches, they trigger these actions twice when working in pull requests.
This is an unnecessary use of actions resources and can slow down
getting work done (e.g. slowed down getting #868 done to unblock a release).

After this change, the github actions in `ci.yaml` will continue to run
on pull request events. It will modify the behavior to only run on push
events when they match one of the following:
- `main` branch: default project branch
- `release/*` branch: branches used to prepare releases
- `snapshot/*` branch: branches used to release one-off snapshots for
  testing
- `v*` tag: release tags.

This also adjusts `go.yml` and `react.yml` to use a consistent set of filters
for running on `push`.

Signed-off-by: Julie Pagano <julie.pagano@chronosphere.io>