[IGNORE] Include retry in dashboard name for tests
I noticed a rare flaky test run related to this in another PR.
We have retry logic set up to catch a rare issue in a test and try again.
When a retry runs on a test set that mutates dashboards, we can hit a
case where the backend throws an error because we try to create a
duplicate dashboard name.

This fix includes the `retry` number in the dashboard name to ensure it
is unique.

Signed-off-by: Julie Pagano <julie.pagano@chronosphere.io>