[IGNORE] Short-term fix for visual test fonts
Prior to this fix, the visual tests were not using the correct font
because the happo asset uploader only caught relative asset links and
our fonts use a full link including localhost.

The addition of the `HAPPO_DOWNLOAD_ALL=true` env var forces the happo
asset uploader to grab all assets, which will grab the fonts. This is
the easiest short-term fix. Their support team will take a look
at adjusting the script to support non-relative asset links with
localhost in them, and we can hopefully pull down and get a fix in the
near future. We have very few external assets, so the side effects of
using this env var should be minimal even if we need it for a longer
period of time.

Signed-off-by: Julie Pagano <julie.pagano@chronosphere.io>