[BUGFIX] Allow editing of time range using inputs (#963)
Prior to this fix, the text inputs for setting a custom time range
attempted to turn the content of the text input into a `Date` and
apply the changes on `change`. This prevented the user from modifying
the input to adjust the value and would prematurely complete the time
range editing interaction.

After this fix, the text inputs are treated as text while being changed.
They are evaluated as dates on blur. This fix also introduces an "apply"
button that must be clicked to submit the changes in the text inputs, so
that the user has control of when the time range editing interaction
completes using text inputs.

This change also fixes a bug where the end time minutes value selected
using the clock face interface was not being saved.

Signed-off-by: Julie Pagano <julie.pagano@chronosphere.io>