[IGNORE] reduce visual test flakiness
Some of the flakiness was likely related to the panel being moused over
as a result of the "trial click" workaround to wait for a stable panel.
Trying out replacing that with a `scrollIntoViewIfNeeded`, which also
checks for a stable element, but does not mouseover, and as an added
benefit, it will make sure the panel is scrolled into view.

https://playwright.dev/docs/actionability

Signed-off-by: Julie Pagano <julie.pagano@chronosphere.io>