[BUGFIX] handle variable default with all in an array
The variable handling assumes that `$__all` is alone as a value, which
was managed in internal interactions, but not on intializing variables,
which can lead to buggy behavior. This commit normalizes the all value
when it is the single value in an array on intialization.

This comment also expands the stories for the relevant components to
duplicate the issue and validate it is fixed.

Signed-off-by: Julie Pagano <julie.pagano@chronosphere.io>