[ENHANCEMENT] always show info icon when panel has a description
Prior to this change, the info icon was only shown when the user was
hovering over the panel, which significantly reduces discoverability
of the description information, especially on larger dashboards.

After this change, the info icon is always shown in "view" mode when
a panel has a description. The hover over the icon to show the
description in a tooltip continues to behave as before.

Signed-off-by: Julie Pagano <julie.pagano@chronosphere.io>