[BUGFIX] miscellaneous fixes to address console errors (#1068)
This commit introduces functionality that fails Playwright-based e2e tests if console errors
are seen during the test because this is often a sign of a subtle bug we would not catch
otherwise. 

As part of introducing this, we need to fix some of the underlying console errors from 
libraries we use (e.g. MUI, React). In one case, we ignore an emotion error because it is
tricky to fix and focused on a use case we are not currently focused on (SSR). Fixes include:

- Fixing React errors related to a text input changing from uncontrolled to controlled because a
  text input could sometimes be `undefined`. These now fall back to `''` instead.
- Fixing a use case where `li` elements were not properly parented to an `ul` or `ol`.
- Fixing a React `key` error by setting `key` value on children in a list.
- Wrapping tooltip children in a `span` to avoid MUI errors when the child is a disabled button.

Signed-off-by: Julie Pagano <julie.pagano@chronosphere.io>