[IGNORE] Add simple typedoc setup (#1056)
I have been struggling to keep track of what Perses exports from the
various UI packages, which can make it tricky to figure out breaking
changes when things are refactored. Typedoc does a nice job evaluating
the entry point and creating documentation, so that seemed like a low
effort way to get this information.

In the short term, storing these as artifacts in github actions runs.
Longer-term, we could potentially publish these in github pages alongside
storybook (good at documenting React components) to provide
documentation for the non-component utilities, types, etc. that Perses
provides.

It is notable that typedoc isn't great at handling cross-package
dependencies at this time, so there are some limitations to what we're
getting. We can watch this issue to see when this is improved.
https://github.com/TypeStrong/typedoc/issues/1835

Signed-off-by: Julie Pagano <julie.pagano@chronosphere.io>