[IGNORE] Add basic stories for time series plugin
This commit adds some basic stories for the following features of the
time series plugin:
- panel options editor
- panel component

These stories are intended to make it easier to develop these features
for developers and to provide consumers with an idea of what these
features will look like.

This commit also fixes a small copy-paste error in one of the storybook
decorators.

Signed-off-by: Julie Pagano <julie.pagano@chronosphere.io>