[BUGFIX] fix date handling in tooltip (#1163)
Previously the date was passed as a localized string (generated in
getNearbySeries() in focused-series.ts) to formatTimeSeriesHeader()
in TooltipContent.tsx, where it was parsed again. However,
`new Date(...)` cannot parse this localized string
(example with German locale: `11. Mai 2023, 6:05:45 PM`).

This commit updates the code to pass the date in milliseconds instead.

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>