[BREAKINGCHANGE] use react-virtuoso for virtualization (#1160)
Previously, the list legend used react-window for virtualization.
We want to switch to using react-virtuoso soon for the table legend
because it has significantly better handling for tables. We do not want
to use two different virtualization libraries, so migrating the list
legend to react-virtuoso in prep for the table legend work.

This should be a no-op from a functionality perspective.

BREAKING CHANGE: Legend no longer takes `listProps` because the associated `initialRowHeight` is no longer needed with the adjusted virtualization implementation.