[BREAKINGCHANGE]: move legend interactions to legend
Prior to this change, much of the legend selection interactions were
managed inside `TimeSeriesChartPanel`, requiring consumers to duplicate
this logic when using the legend in other use cases.

After this change, the legend selection interactions are managed inside
`Legend` and its child components, so that it can be used wherever the
legend is. The selection state remains "controlled."

BREAKINGCHANGE: Refactoring `Legend` to manage legend interactions
includes the following breaking changes:
- The `Legend` component now requires the following props to control the
  selected state of the items: `selectedItems`, `onSelectedItemsChange`.
- The `LegendItem` type no longer allows the `isSelected` property. The
  selected state of items are now managed by the `selectedItems` prop on
  the `Legend` component.
- The `onClick` property on the `LegendItem` type is now optional because
  it is no longer used to manage the selected state of items. Instead
  use the `onSelectedItemsChange` prop on the `Legend` component if you
  need to listen for selection changes.

Signed-off-by: Julie Pagano <julie.pagano@chronosphere.io>