[BREAKINGCHANGE] Adjust MUI bg tokens to simplify theming (#1222)
* [BREAKINGCHANGE] Adjust MUI bg tokens to simplify theming

Prior to this change, chunks of the codebase would have to switch
between `paper` and `default` depending on if the theme was in `light`
or `dark` mode. After this change, consistent values can be used for
each.

BREAKINGCHANGE: This change modifies the MUI theme design tokens for
the `paper` and `default` background colors and some related components
that use these tokens. Consumers may need to make some associated
theming changes to handle this.

Signed-off-by: Julie Pagano <julie.pagano@chronosphere.io>

* Fix table

Signed-off-by: Julie Pagano <julie.pagano@chronosphere.io>

* Fix license headers

Signed-off-by: Julie Pagano <julie.pagano@chronosphere.io>

* Fix table hover

Signed-off-by: Julie Pagano <julie.pagano@chronosphere.io>

---------

Signed-off-by: Julie Pagano <julie.pagano@chronosphere.io>