[BREAKINGCHANGE] reorg legend code to separate plugin spec & component concerns (#1223)
The purpose of this change is to improve the separation of concerns between the `Legend` component and legend-related behavior in panel plugin specifications. 

BREAKINGCHANGE: this code reorganizes some types, plugin system components, and utilities related to legends to better organize them. You may need to change the package you import from for some of these values. 
- Legend code related to the `Legend` component continues to live in the `components` package.
- Legend code related to panel specs moves to the `plugin-system` package.
- Common legend code lives in `core`.

Signed-off-by: Julie Pagano <julie.pagano@chronosphere.io>