[BREAKINGCHANGE] Improve calculations for gauge and stat charts
This change includes a handful of improvements for the calculation
setting currently used in the gauge and stat charts and the associated
utility functions for generating those calculations:
- Gauge and stat charts now communicate when the calculation is `null`.
  Previously, they would show a NaN in these cases, which is confusing.
- Calculation for `Mean` is now labeled as `Average` for slightly better
  usability/readablity. The underlying constant remains the same.
- Calculation utilities now can return `null` in cases where the data
  is `null`. Previously, they would return `NaN`.
- Adjusted several utility functions to handle the fact that values can
  now be `null`.
- Migration code now translates `firstNotNull` to `FirstNumber`, which
  is more equivalent than the previous translation to `First`.
- Added `FirstNumber` calculation option to provide users the ability to
  calculate the first non-null value.

BREAKINGCHANGE: the return values for calculation functions now return
`null` instead of `NaN` when values are `null`. Several type signatures
for working with values have been modified to include `null` values and
may need to be accounted for.

Signed-off-by: Julie Pagano <julie.pagano@chronosphere.io>