Implement the `healthcheck` method for Prometheus datasources (#1595)
The `DatasourceClient` interface defines a `healthcheck` method, which
hasn't actually been implemented. This implements it for the `PrometheusDatasourceClient`
as a first step towards being able to test the health of a datasource.

The actual implementation calls the [/-/healthy](https://prometheus.io/docs/prometheus/latest/management_api/#health-check)
endpoint in the Prometheus API, rather than running a query. I figured
this was a better way to test the health rather than running a dummy
query and checking the result, as running a query is a more expensive
and subjects us to the vagaries of PromQL, and potential
query timeouts which aren't directly related to whether the datasource
is configured correctly.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>