Prevent using dashboard datasources before they're saved
Previously, when a dashboard level datasource is added to a dashboard,
it is immediately available for use in the dashboard. This causes an
error because the server side (e.g. the proxy endpoints) isn't created
until the dashboard is saved.

Here we introduce the idea of "saved" datasources, which are
the datasources that are available for use in the dashboard because they
have been created through a dashboard save.

We then use this list of saved datasources to disable entries in the
datasource select dropdown, with an indication that the dashboard needs
to be saved before the datasource can be used.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>