[ENHANCEMENT] Prevent using dashboard datasources before they're saved (#1603)
* Prevent using dashboard datasources before they're saved

Previously, when a dashboard level datasource is added to a dashboard,
it is immediately available for use in the dashboard. This causes an
error because the server side (e.g. the proxy endpoints) isn't created
until the dashboard is saved.

Here we introduce the idea of "saved" datasources, which are
the datasources that are available for use in the dashboard because they
have been created through a dashboard save.

We then use this list of saved datasources to disable entries in the
datasource select dropdown, with an indication that the dashboard needs
to be saved before the datasource can be used.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>

* Allow some new / updated datasources to be used

This updates the `EditDatasourcesButton` to allow some new or updated datasources
to be used. In particular, datasources that are either direct, or are
proxied but have not been updated are allowed to be used immediately
after being created or updated.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>

---------

Signed-off-by: sinkingpoint <colin@quirl.co.nz>