Make Datasource Saving/Updating Async
In order to run tests, we need to know when the datasource has been
saved (mainly for proxy datasources that need the server side setup).
This commit goes through all the paths that save a datasource
and makes them async, returning a promise that resolves when the
datasource has been saved.

Among the various changes, we have to move a couple of methods from old
style `Promise`s to `async`/`await` style, and we have to change the
mutations from `mutate` to `mutateAsync` to ensure that the promise
actually represents the mutation.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>