Test datasources when they're saved
This updates the datasource editor to test the datasource when it's
saved. This pretty much mirrors how Grafana does this, with a "Save and Test"
button when editing datasources that saves the datasource and then runs
a test. We use the previously created healthcheck, and the fact
that saves are now asynchronous to run the check after the save has
completed.

There's a few things of note:

- Users of the <DatasourceEditorForm /> component no longer need to call
their `onClose` when the datasource is saved - the editor will call
it when the health check is sucessful.

- Dashboard local datasources _are not healthchecked_, for mostly the
same reasons as why they can't be immediately used - they don't exist
serverside until the dashboard is saved.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>