Wrap PrometheusDatasourceEditor with react-hook-form Controllers
Previously, the fields created with the PrometheusDatasourceEditor weren't
wired into the form infrastructure. This means that, among other things,
if you edit the URL of the datasource, the form doesn't know that the
datasource has changed and doesn't allow you to save.

This change wraps the PrometheusDatasourceEditor inputs with react-hook-form
Controllers, which means that the form is now aware of changes.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>