[ENHANCEMENT] Add in a proxy for unsaved datasources (#1635)
* Add in a proxy for unsaved datasources

This adds in three new API endpoints on the proxy, that allow requesting
a datasource that has not been saved to the database. These methods instead
take a POST request with the datasource definition, request method, and request
body in the body and uses that to construct the request to proxy.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>

* Authenticate the unsaved endpoints

This commit adds authentication to the unsaved endpoints, in order
to prevent this being an Open Proxy. In particular, we authenticate
against the permissions to Create a datasource (which is generally what
this request would precede).

Signed-off-by: sinkingpoint <colin@quirl.co.nz>

---------

Signed-off-by: sinkingpoint <colin@quirl.co.nz>