Move Migrate Button into "Add Dashboard" (#1726)
This introduces a "Button Menu", a CRUD Button that has a drop down next
to it for more options, not unlike Githubs Merge button etc. We then
use that component to move the Migrate button from the header
into an option below the Add Dashboard button on the home page.

This tidies up the UI a bit, but also sets the stage for a more generic
"Import Dashboard" button that can be used to import dashboards from
other sources.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>