[FEATURE] Allow importing Perses dashboards (#1769)
* Rename "Migrate Dashboard" -> "Import Dashboard"

This renames the Migrate Dashboard button to prepare
for the new workflow of importing perses dashboards.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>

* Allow importing Perses dashboads in the UI

This splits the MigrateView into two separate flows, one for Grafana and
one for Perses. The Grafana flow is unchanged, but the Perses flow
skips the migrate translation step, going straight to the
project select and import.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>

* Rename Migrate -> Import

This renames all the migrate paths and components to import
to reflect that /import can now be used to import both
Grafana dashboards (the old migrate flow), and Perses dashboards.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>

---------

Signed-off-by: sinkingpoint <colin@quirl.co.nz>