Backend work for usage tracking
Here we add the backend side of usage tracking. We
introduce a new view service that will be used to
track the usage of the API, bound to the /api/v1/view endpoint.

We then add a new implementation of that service
that tracks views in Prometheus metrics, with the
flexibility to swap out the implementation later.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>