[BUGFIX] replace unstable JSON.stringify() with custom string in <PluginKindSelect>
I'm seeing this error on the Explore page in the JS console on Firefox:
```
MUI: You have provided an out-of-range value `{"kind":"PrometheusTimeSeriesQuery","type":"TimeSeriesQuery"}` for the select component.
Consider providing a value that matches one of the available options or ''.
The available values are `{"type":"TimeSeriesQuery","kind":"PrometheusTimeSeriesQuery"}`.
```

This is because the order of keys in Objects is undefined, therefore
the output of JSON.stringify() is not guaranteed to be the same for two
objects with the same keys and values.

I followed the same approach as for DatasourceSelect:
https://github.com/perses/perses/blob/cf87ec4b6bd474f2d58c01709a36ff37bcde9512/ui/plugin-system/src/components/DatasourceSelect.tsx#L139-L167

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>