[BUGFIX] fix React warning when using TraceQL Editor (#2031)
* [BUGFIX] fix React warning when using TraceQL Editor

Fixes the following React warning:
```
Warning: React does not recognize the `completeConfig` prop on a DOM element. If you intentionally want it to appear in the DOM as a custom attribute, spell it as lowercase `completeconfig` instead. If you accidentally passed it from a parent component, remove it from the DOM element.
```

The completeConfig is required for the PromQLExtension, however for
TraceQL no such extension exists at the moment.

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>

* run linter

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>

---------

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>