[BUGFIX] explore page: sync query parameters to explorer state
Previously, clicking on a `<Link>` (without `reloadDocument` attribute) to the
explore page with queries in the URL did update the query parameters,
but the updated state was not reflected in the `ExplorerManagerProvider`.
This is because the query parameters were only considered for the initial
state of the `ExplorerManagerProvider`, but other changes of the query
parameters were not considered until a full page refresh.

Changes:
* switched the `explorer` type from a number to a string to improve
  the permalink, i.e. remove the magic number
* keep the URL changes in the browser history, to allow easy navigation from
  e.g the trace table to gantt chart and back via the browser's back button
* refactored the `ExplorerManagerProvider` to keep the state
  in a single place (state based on the query parameters)

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>