[BUGFIX] tempo query: disable syntax highlighting if query is a trace id (#2242)
In the explore view, the input field accept either a TraceQL query or a Trace ID,
and displays the trace table or the Gantt chart based on the value.
Unfortunately, the lezer grammar (correctly) marks Trace IDs as invalid TraceQL queries.
Therefore, let's disable syntax highlighting if the input is a Trace ID.

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>