[BUGFIX] tempo: support search results with mixed vParquet3 and vParquet4 blocks
When upgrading to Tempo v2.6.0, the vParquet4 block is the default,
which includes the `serviceStats` field in the search API response.

The previously stored blocks (using vParquet3 format or older) do not
contain this field yet, therefore it's possible to receive a search API
response where some traces contain the `serviceStats` field and some
traces don't contain this field.

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>