[IGNORE] Fix flaky StatChart panel e2e test (#2790)
Due to the size of the StatChart panel in the example dashboard, the
edit button was in the action menu. Unfortunately locating elements in
the MaterialUI <Popover> element is unstable at times, therefore let's
just increase the StatChart panel width for the e2e test, so the edit
button is directly in the Panel header instead of the MUI <Popover>
element.

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>