[IGNORE] remove ability to save user data with k8s auth
Remove the need for hashing the file names and ignoring the validation
on the user by instead just not saving information onto the perses
backend. The jwt token is still set with the username for the perses
frontend to use. All user, role, and rolebinding endpoints are disabled
when k8s auth is enabled except for /users/{username}/permissions

Signed-off-by: PeterYurkovich <pyurkovi@redhat.com>