Add Execute Query Button to query editors (#3049)
Currently when you edit a query it runs whenever there is a re-render done (generally
when you click off of the editor box). This leads to a weird experience if you're not used to that,
where your query seemingly runs at random. Here we add a new button to Query Editors
that gives users the ability to manually run their query, making this flow more explicit.

Signed-off-by: sinkingpoint <colin@quirl.co.nz>