[IGNORE] add panelOptions.showIcons setting (#3244)
Currently, we show most of the panel icons only on hover, unless the
dashboard is in editing mode, or the panel is in fullscreen mode.

In the Tempo explore view, the panel is in fullscreen mode, but doesn't
have readHandlers.isPanelViewed set (because it's always fullscreen,
there is no option to change the panel size).

This PR adds an explicit `showIcons` setting which can be either `hover`
or `always`, and by default uses the current behavior (`always` if the
dashboard is in editing mode or panel is in fullscreen mode, `hover`
otherwise).

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>