[BUGFIX] Use nonroot instead of nobody
Running a process like Perses -- which does not require privileges -- as
`nobody` is strictly better than running as root. The Prometheus
Dockerfile does the same [1].

However, the original purpose of the `nobody` user was in the context of
NFS [2]: Files that were created by `root` on the NFS client are owned
by `nobody` on the NFS server. For that reason, running a process (such
as Perses) as `nobody` might be dangerous if the host is also an NFS
server: If Perses somehow manages to access an NFS export, it can access
all remote files owned by the NFS clients' `root` user [3].

The distroless images, which Perses uses, have a dedicated user for
running unprivileged services: The nonroot user. This commit switches to
distroless images which have that user as a default.

[1]: https://github.com/prometheus/prometheus/blob/f6ca7145ca2ffe8bdd81e373657c740544abc5ac/Dockerfile
[2]: https://0xjet.github.io/3OHA/2022/06/01/post.html
[3]: https://unix.stackexchange.com/q/9840/55203

Signed-off-by: runiq <hey@runiq.de>